Objects exported from other packages

These objects are imported from other packages. Follow the links below to see their documentation.

data.table

%between%, %chin%, %inrange%, %like%, .BY, .GRP, .I, .N, .SD, :=, address, alloc.col, as.data.table, as.IDate, as.ITime, between, chgroup, chmatch, chorder, CJ, copy, cube, data.table, fcoalesce, fifelse, fintersect, first, foverlaps, frank, frankv, fread, frollapply, frollmean, frollsum, fsetdiff, fsetequal, fsort, funion, fwrite, getDTthreads, getNumericRounding, groupingsets, haskey, hour, IDateTime, indices, inrange, is.data.table, isoweek, key, key<-, key2, last, like, mday, minute, month, nafill, quarter, rbindlist, rleid, rleidv, rollup, rowid, rowidv, second, set, set2key, set2keyv, setattr, setcolorder, setDF, setDT, setDTthreads, setindex, setindexv, setkey, setkeyv, setnafill, setnames, setNumericRounding, setorder, setorderv, shift, shouldPrint, SJ, tables, test.data.table, timetaken, transpose, truelength, tstrsplit, uniqueN, wday, week, yday, year

maditr

%>%

magrittr

%$%, %<>%, %>%, %T>%